POKEMON-26 The bottom navigation bar works on app and load other scarfold pages on it
