+++ frontend/pokemoney/lib/RouteGenerator.dart
@@ -1,14 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:pokemoney/app.dart';
 import 'package:pokemoney/pages/barrel.dart';
 
 class RouteGenerator {
-  static const String homePage = '/';
+  static const String app = '/';
+  static const String homePage = '/home';
+  static const String fundsPage = '/funds';
+  static const String ledgerBooksPage = '/ledgerBooks';
+  static const String accountsPage = '/accounts';
   static const String loginPage = '/login';
   static const String signUpPage = '/signUp';
-  static const String forgotPasswordPage = 'forgotPassword';
-  static const String accountsPage = '/accounts';
-  static const String ledgerBooksPage = '/ledgerBooks';
-  static const String fundsPage = '/funds';
+  static const String forgotPasswordPage = '/forgotPassword';
   static const String settingsPage = '/settings';
   static const String statsPage = '/stats';
   static const String stylePage = '/style';
@@ -18,10 +20,27 @@ class RouteGenerator {
   RouteGenerator._() {}
 
   static Route<dynamic> generateRoute(RouteSettings settings) {
+    
     switch (settings.name) {
+      case app:
+        return MaterialPageRoute(
+          builder: (_) => App(),
+        );
       case homePage:
         return MaterialPageRoute(
-          builder: (_) => const HomePage(),
+          builder: (_) => HomePage(),
+        );
+      case fundsPage:
+        return MaterialPageRoute(
+          builder: (_) => FundsPage(),
+        );
+      case ledgerBooksPage:
+        return MaterialPageRoute(
+          builder: (_) => LedgerBooksPage(),
+        );
+      case accountsPage:
+        return MaterialPageRoute(
+          builder: (_) => AccountsPage(),
         );
       case loginPage:
         return MaterialPageRoute(
@@ -34,18 +53,6 @@ class RouteGenerator {
       case forgotPasswordPage:
         return MaterialPageRoute(
           builder: (_) => ForgotPasswordPage(),
-        );
-      case accountsPage:
-        return MaterialPageRoute(
-          builder: (_) => AccountsPage(),
-        );
-      case ledgerBooksPage:
-        return MaterialPageRoute(
-          builder: (_) => LedgerBooksPage(),
-        );
-      case fundsPage:
-        return MaterialPageRoute(
-          builder: (_) => FundsPage(),
         );
       case settingsPage:
         return MaterialPageRoute(
