+++ frontend/pokemoney/lib/pages/FundsPage.dart
@@ -7,12 +7,6 @@ class FundsPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     // TODO: implement build
-    return Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        title: Text("Funds"),
-        backgroundColor: AppColors.surfaceContainer,
-      ),
-    );
+    return Scaffold();
   }
 }
