+++ frontend/pokemoney/lib/pages/HomePage.dart
@@ -3,81 +3,13 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:pokemoney/constants/AppColors.dart';
-
 import 'package:pokemoney/pages/barrel.dart';
 
-class HomePage extends StatefulWidget {
-  const HomePage({super.key});
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  int _currentIndex = 0;
-  static List<String> _routes = [
-    RouteGenerator.homePage,
-    RouteGenerator.fundsPage,
-    RouteGenerator.ledgerBooksPage,
-    RouteGenerator.accountsPage
-  ];
-
-  void _onItemTapped(int index) {
-    setState(() {
-      _currentIndex = index;
-      Navigator.of(context)?.pushNamed(_routes[_currentIndex]);
-    });
-  }
+class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: AppColors.surface,
-      appBar: AppBar(
-          iconTheme: IconThemeData(color: AppColors.textPrimary),
-          backgroundColor: AppColors.surfaceContainer,
-          actions: [
-            IconButton(
-              icon: const Icon(
-                FontAwesomeIcons.bell,
-                color: AppColors.textPrimary,
-              ),
-              tooltip: 'Show Snackbar',
-              onPressed: () {
-                ScaffoldMessenger.of(context).showSnackBar(
-                    const SnackBar(content: Text('This is a notifications')));
-              },
-            ),
-          ]),
-      drawer: CustomNavBar(),
-      bottomNavigationBar: BottomNavigationBar(
-        type: BottomNavigationBarType.fixed,
-        backgroundColor: AppColors.surfaceContainer,
-        selectedItemColor: AppColors.textPrimary,
-        unselectedItemColor: AppColors.textSecondary,
-        selectedFontSize: 14,
-        unselectedFontSize: 14,
-        currentIndex: _currentIndex,
-        onTap: _onItemTapped,
-        items: [
-          BottomNavigationBarItem(
-            label: 'Home',
-            icon: Icon(Icons.home),
-          ),
-          BottomNavigationBarItem(
-            label: 'Funds',
-            icon: Icon(Icons.credit_card),
-          ),
-          BottomNavigationBarItem(
-            label: 'Bills',
-            icon: Icon(FontAwesomeIcons.fileInvoiceDollar),
-          ),
-          BottomNavigationBarItem(
-            label: 'Accounts',
-            icon: Icon(Icons.person),
-          ),
-        ],
-      ),
-    );
+    return Center(child: Text('Home Page'));
   }
 }
