+++ frontend/pokemoney/lib/pages/barrel.dart
@@ -11,4 +11,4 @@ export 'StylePage.dart';
 export 'SettingsPage.dart';
 export 'PrivacySecurityPage.dart';
 export 'HelpPage.dart';
-export 'RouteGenerator.dart';
+export 'package:pokemoney/RouteGenerator.dart';
